Proof of concept for APL UX.
This commit is a proof of concept of APL based experience. It captures how we
can show different APL documents while rendering the results of previous question
and rendering the next question. We used auto generated 'UserEvents' to obtain this experience.

Quite a convoluted way of obtaining desired experience. Tried 'progressive responses'
and a few other techniques but none of them worked.

For future reference, this commit is not a great stopping point. I will clean this up
in the next few commits.