Implemented an independent APL experience.
With this commit, it is possible to play the game completely through
APL without having to use voice (beyond launching the skill).

To accomplish this, the results delivery and next question deliver were
decoupled. Once the user clicks on an option, we just deliver the results
and schedule an auto generated event. Once the results are read out, the
auto generated event is fired and we deliver the next question in response
to this event.